+++ lib/features/family/features/box_origin/box_origin_selection_page.dart
@@ -1,24 +1,17 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:givt_app/app/injection/injection.dart';
 import 'package:givt_app/core/enums/amplitude_events.dart';
 import 'package:givt_app/core/enums/collect_group_type.dart';
-import 'package:givt_app/core/enums/country.dart';
-import 'package:givt_app/features/auth/cubit/auth_cubit.dart';
 import 'package:givt_app/features/family/app/family_pages.dart';
 import 'package:givt_app/features/family/features/impact_groups/widgets/dialogs/box_origin_outcome_dialog.dart';
+import 'package:givt_app/features/family/features/parent_giving_flow/presentation/pages/organisation_list_family_page.dart';
 import 'package:givt_app/features/family/shared/design/components/components.dart';
-import 'package:givt_app/features/family/shared/widgets/buttons/givt_back_button_flat.dart';
-import 'package:givt_app/features/family/shared/widgets/inputs/family_search_field.dart';
-import 'package:givt_app/features/family/shared/widgets/loading/custom_progress_indicator.dart';
 import 'package:givt_app/features/family/shared/widgets/texts/shared_texts.dart';
 import 'package:givt_app/features/family/utils/family_app_theme.dart';
 import 'package:givt_app/features/give/bloc/bloc.dart';
-import 'package:givt_app/l10n/l10n.dart';
 import 'package:givt_app/shared/models/analytics_event.dart';
-import 'package:givt_app/shared/widgets/fun_scaffold.dart';
 import 'package:givt_app/utils/analytics_helper.dart';
 
 class BoxOrignSelectionPage extends StatefulWidget {
@@ -40,16 +33,6 @@ class _BoxOrignSelectionPageState extends State<BoxOrignSelectionPage> {
   bool isLoading = false;
 
   @override
-  void initState() {
-    super.initState();
-    getIt<OrganisationBloc>().add(
-      OrganisationFetchForSelection(
-        Country.fromCode(context.read<AuthCubit>().state.user.country),
-      ),
-    );
-  }
-
-  @override
   void dispose() {
     controller.dispose();
     super.dispose();
@@ -57,73 +40,19 @@ class _BoxOrignSelectionPageState extends State<BoxOrignSelectionPage> {
 
   @override
   Widget build(BuildContext context) {
-    final locals = context.l10n;
-    return FunScaffold(
-      appBar: const FunTopAppBar(
-        leading: GivtBackButtonFlat(),
-        title: 'Where did you get it?',
-      ),
-      body: BlocConsumer<OrganisationBloc, OrganisationState>(
-        bloc: bloc,
-        listener: (context, state) {
-          if (state.status == OrganisationStatus.error) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                content: Text(locals.somethingWentWrong),
-              ),
-            );
-          }
-        },
-        builder: (context, state) {
-          return Column(
-            children: [
-              FunOrganisationFilterTilesBar(
-                stratPadding: 0,
-                removedTypes: [
-                  CollectGroupType.campaign.name,
-                  CollectGroupType.artists.name
-                ],
-              ),
-              const SizedBox(height: 16),
-              FamilySearchField(
-                autocorrect: false,
-                controller: controller,
-                onChanged: (value) =>
-                    bloc.add(OrganisationFilterQueryChanged(value)),
-              ),
-              const SizedBox(height: 8),
-              if (state.status == OrganisationStatus.filtered)
-                Expanded(
-                  child: ListView.separated(
-                    separatorBuilder: (_, index) => const Divider(
-                      height: 0.1,
-                      color: FamilyAppTheme.neutralVariant95,
-                    ),
-                    shrinkWrap: true,
-                    itemCount: state.filteredOrganisations.length,
-                    itemBuilder: (context, index) {
-                      return _buildListTile(
-                        type: state.filteredOrganisations[index].type,
-                        title: state.filteredOrganisations[index].orgName,
-                        isSelected: selectedIndex == index,
-                        onTap: () {
-                          setState(() {
-                            selectedIndex = index;
-                          });
-                        },
-                      );
-                    },
-                  ),
-                )
-              else
-                const Center(
-                  child: CustomCircularProgressIndicator(),
-                ),
-            ],
-          );
-        },
-      ),
-      floatingActionButton: FunButton(
+    return OrganisationListFamilyPage(
+      onTap: (collectgroup) {
+        final index = bloc.state.filteredOrganisations.indexOf(collectgroup);
+        setState(() {
+          selectedIndex = index;
+        });
+      },
+      title: 'Where did you get it?',
+      removedCollectGroupTypes: const [
+        CollectGroupType.campaign,
+        CollectGroupType.artists,
+      ],
+      fab: FunButton(
         isDisabled: selectedIndex == -1,
         isLoading: isLoading,
         text: 'Confirm',
